refactor: move heading logic into TableOfContents
